pyproject.toml
[project]
name = "ppaa-experiments"
version = "0.0.0"
description = "Compare post-processing aa methods (and ssaa)"
authors = [{ name = "Almar Klein" }]
dependencies = [
    "wgpu",
    "jinja2",
    "numpy",
    "pillow",
    "aggdraw",
    "matplotlib",
    "scipy",
    "ruff",
]

[project.urls]
Homepage = "https://github.com/almarklein/ppaa-experiments"


[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["F", "E", "W", "B", "RUF"]
ignore = [
    "E501",   # Line too long
    "B007",   # Loop control variable `x` not used within loop body
    "RUF012", # Mutable class attributes should be annotated with `typing.ClassVar`
]
